# projet-08-Ot-gether-back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ot'gether API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/register": {
            "post": {
                "summary": "Register a new user",
                "operationId": "8e4e3cb7b46126f4696379d3c7eeb4ad",
                "parameters": [
                    {
                        "name": "lastName",
                        "in": "query",
                        "description": "User's lastname",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "firstName",
                        "in": "query",
                        "description": "User's firstname",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "pseudo",
                        "in": "query",
                        "description": "User's pseudo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "description": "User's city",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User's email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User's password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "confirmPassword",
                        "in": "query",
                        "description": "User's confirm password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/check-email": {
            "post": {
                "summary": "Verify if an email is already used",
                "operationId": "0a37eff402242460d34b78324e6c75e3",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User's email",
                        "required": true,
                        "schema": {
                            "type": "email"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "isUsed"
                    },
                    "422": {
                        "description": "Errors"
                    }
                }
            }
        },
        "/api/categories": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Get list of categories",
                "operationId": "ac9c4d978ea4e5b72915ca9728e5b8b1",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Register a new category",
                "operationId": "cca49a87d9b0ebb9a98ad7c154b71737",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Category's name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Category created successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/categories/{category}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Get a category by id",
                "operationId": "ff3ab984d3fe877eb16da6009394d694",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "ID of category to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Category"
                ],
                "summary": "Delete a category by id",
                "operationId": "88c22abd8149aea4b3bdcb280a0fddf7",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "ID of category to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Category"
                ],
                "summary": "Update a category by id",
                "operationId": "5afb9969f22296ff47a6d827a9028b9c",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "ID of category to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Category name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/events/{event}/messages": {
            "get": {
                "tags": [
                    "Chat"
                ],
                "summary": "Get all messages from an event",
                "operationId": "a4cb12c211432ae272e4a77641c1b84d",
                "parameters": [
                    {
                        "name": "event",
                        "in": "query",
                        "description": "ID of event to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of messages"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Chat"
                ],
                "summary": "Send a message to an event",
                "operationId": "6a75cca5a67e711561da5a37dad38b36",
                "parameters": [
                    {
                        "name": "eventId",
                        "in": "path",
                        "description": "ID of event to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "content",
                        "in": "query",
                        "description": "Content of the message",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message sent"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}